Update ModelConfidence.md
Completed this chapter. Two sentences I would like you to review:

1st: However, in most applied cases you will not be able to apply this normality assumption.   (I wanted to use the word/variation of 'apply' only once but couldn't figure out how to do without changing meaning.)


2nd: under the section labeled:  Confidence and Philosophy

The quality of a model in the eye of an evaluator is significantly influenced by the world-view (more or less coherent and the philosophical orientation [if any] of the evaluator.)

I changed to read: 
From an evaluator’s view, the quality of a model is significantly influenced by the evaluator’s world-view and philosophical orientation (if any).